third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/getComputedStyle-list-style-shorthand-expected.txt

Tests that the list-style shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('list-style') is "outside url(\"dummy://test.png\") circle"
PASS computedStyle.getPropertyValue('list-style') is "outside none disc"
PASS computedStyle.getPropertyValue('list-style') is "outside none circle"
PASS computedStyle.getPropertyValue('list-style') is "outside none none"
PASS computedStyle.getPropertyValue('list-style') is "outside url(\"dummy://test.png\") circle"
PASS successfullyParsed is true

TEST COMPLETE
